refactor(test): extract shared long-test harness
